+++ src/v1/BasicTypes/Functions/isFunction.spec.ts
@@ -31,37 +31,38 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { FIND_PROPERTY_NAMES_DEFAULT_OPTIONS, findMethodNames } from "./Filters";
-import { FIND_PROPERTIES_FILTER_DROP_CONSTRUCTORS } from "./Filters/defaults/FIND_PROPERTIES_FILTER_DROP_CONSTRUCTORS";
-import { FIND_PROPERTIES_FILTER_DROP_INTERNAL } from "./Filters/defaults/FIND_PROPERTIES_FILTER_DROP_INTERNAL";
 
-/**
- * `getPublicMethodNames()` is a data filter. It returns a list of all
- * methods that form the object's public API.
- *
- * - all methods defined on `target` and its base classes (inc
- *   Object.prototype)
- * - that aren't constructors, and
- * - that don't start with an underscore (ie suggest they're protected
- *   or private)
- *
- * Getters and Setters are NOT treated as public methods.
- *
- * @param target
- * The object to inspect.
- * @returns
- * A list of all method names that exist on the object instance. Order of
- * the list is not guaranteed. The list will not contain duplicates.
- *
- * @category BasicTypes
- */
-export function getPublicMethodNames(
-    target: object
-): string[] {
-    return findMethodNames(
-        target,
-        FIND_PROPERTY_NAMES_DEFAULT_OPTIONS,
-        FIND_PROPERTIES_FILTER_DROP_CONSTRUCTORS,
-        FIND_PROPERTIES_FILTER_DROP_INTERNAL,
-    );
-}
+import { describe } from "mocha";
+import { expect } from "chai";
+import { isFunction } from "./isFunction";
+
+describe("isFunction()", () => {
+    it("returns `true` when given a function", () => {
+        const inputValue: unknown = () => true;
+        const expectedValue = true;
+
+        const actualValue = isFunction(inputValue);
+        expect(actualValue).to.equal(expectedValue);
+
+        // belt n braces ... won't compile if isFunction()
+        // returns the wrong type
+        if (isFunction(inputValue)) {
+            inputValue();
+        }
+    });
+
+    it("returns `false` otherwise", () => {
+        [
+            null,
+            [ 1, 2, 3, 4, 5, ],
+            true,
+            false,
+            100,
+            100.101,
+            {},
+            "hello world"
+        ].forEach((val) => {
+            expect(isFunction(val)).to.equal(false, "failed on " + val);
+        })
+    })
+});