+++ src/v1/Errors/ObjectCannotBeEmpty/ObjectCannotBeEmptyError.ts
@@ -31,37 +31,29 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { FIND_PROPERTY_NAMES_DEFAULT_OPTIONS, findMethodNames } from "./Filters";
-import { FIND_PROPERTIES_FILTER_DROP_CONSTRUCTORS } from "./Filters/defaults/FIND_PROPERTIES_FILTER_DROP_CONSTRUCTORS";
-import { FIND_PROPERTIES_FILTER_DROP_INTERNAL } from "./Filters/defaults/FIND_PROPERTIES_FILTER_DROP_INTERNAL";
+import { AppError, AppErrorData, makeStructuredProblemReport } from "../../ErrorHandling";
+import { makeHttpStatusCode } from "../../SupportingTypes";
+import { MODULE_NAME } from "../defaults/MODULE_NAME";
+import { ObjectCannotBeEmptyData } from "./ObjectCannotBeEmptyData";
 
 /**
- * `getPublicMethodNames()` is a data filter. It returns a list of all
- * methods that form the object's public API.
+ * `ObjectCannotBeEmptyError` is a throwable Error. It is thrown whenever
+ * we unexpectedly encounter an object that has no properties.
  *
- * - all methods defined on `target` and its base classes (inc
- *   Object.prototype)
- * - that aren't constructors, and
- * - that don't start with an underscore (ie suggest they're protected
- *   or private)
- *
- * Getters and Setters are NOT treated as public methods.
- *
- * @param target
- * The object to inspect.
- * @returns
- * A list of all method names that exist on the object instance. Order of
- * the list is not guaranteed. The list will not contain duplicates.
- *
- * @category BasicTypes
+ * @category Errors
  */
-export function getPublicMethodNames(
-    target: object
-): string[] {
-    return findMethodNames(
-        target,
-        FIND_PROPERTY_NAMES_DEFAULT_OPTIONS,
-        FIND_PROPERTIES_FILTER_DROP_CONSTRUCTORS,
-        FIND_PROPERTIES_FILTER_DROP_INTERNAL,
-    );
+export class ObjectCannotBeEmptyError extends AppError<ObjectCannotBeEmptyData> {
+    public constructor(params: ObjectCannotBeEmptyData & AppErrorData) {
+        const spr = makeStructuredProblemReport<ObjectCannotBeEmptyData>({
+            definedBy: MODULE_NAME,
+            description: "empty object is not supported here",
+            errorId: params.errorId,
+            extra: {
+                public: params.public
+            },
+            status: makeHttpStatusCode(422),
+        });
+
+        super(spr);
+    }
 }